+++ src/main/java/org/folio/am/domain/entity/ModuleDiscoveryEntity.java
@@ -5,13 +5,13 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
-import org.hibernate.annotations.Where;
+import org.hibernate.annotations.SQLRestriction;
 
 @Data
 @Entity
 @Table(name = "module")
 @EqualsAndHashCode(callSuper = true)
-@Where(clause = "discovery_url IS NOT NULL")
+@SQLRestriction("discovery_url IS NOT NULL")
 public class ModuleDiscoveryEntity extends ArtifactEntity {
 
   @Column(name = "discovery_url")